+++ CHANGELOG.md
@@ -2,6 +2,26 @@
 
 All notable changes to this project will be documented in this file.
 
+## [0.57.0] - 2023-05-10
+
+### New
+
+- Case-insensitive filtering in args:
+  `blockchain.block(hash)`, `blockchain.transaction(hash)`, `blockchain.message(hash)`
+
+- Address filtering in args (accepts address in any format):
+  `blockchain.account(address)`,  `messages.src`, `messages.dst`, `transactions.account_addr`, `accounts.id`
+
+- Address formatting args in fields:
+  `blockchain.account.info.address`, `blockchain.transaction.account_addr`,
+  `blockchain.message.src`, `blockchain.message.dst`,
+  `messages.src`, `messages.dst`, `transactions.account_addr`, `accounts.id`
+
+- `accounts(filter:{id:{eq:""}})`, `accounts(filter:{id:{in:[""]}})`, `blockchain.account(address:"")`
+  returns account of type `NonExist` if an account is missing in current shard state.
+
+- Support new config parameter 44 (black list)
+
 ## [0.56.0] - 2023-05-09
 
 ### New